+++ apps/client/src/app/pages/admin/admin-page-routing.module.ts
@@ -3,8 +3,8 @@ import { RouterModule, Routes } from '@angular/router';
 import { AdminJobsComponent } from '@ghostfolio/client/components/admin-jobs/admin-jobs.component';
 import { AdminMarketDataComponent } from '@ghostfolio/client/components/admin-market-data/admin-market-data.component';
 import { AdminOverviewComponent } from '@ghostfolio/client/components/admin-overview/admin-overview.component';
+import { AdminPlatformComponent } from '@ghostfolio/client/components/admin-platform/platform.component';
 import { AdminUsersComponent } from '@ghostfolio/client/components/admin-users/admin-users.component';
-import { AdminPlatformComponent } from '@ghostfolio/client/components/platform/platform.component';
 import { AuthGuard } from '@ghostfolio/client/core/auth.guard';
 
 import { AdminPageComponent } from './admin-page.component';